# Use precompiled regex for ID extractor