When using regular expressions with the ID extractor, we compile them on every request, following the suggestions from the performance document, this modifies the code in order to compile the expression only once.

To test this, the following manifest can be used:
```json
  "custom_middleware": {
    "auth_check": {
      "name": "MyAuthCheck"
    },
    "id_extractor": {
      "extract_from": "header",
      "extract_with": "regex",
      "extractor_config": {
        "header_name": "Authorization",
        "regex_expression": "prefix(.*)",
        "regex_match_index": 0
      }
    },
    "driver": "grpc"
  }
```