@matiasinsaurralde 2018-04-30
@dencoded I have refactored it a bit in order to re-use the decoded structure.