@buger 2018-05-11
@matiasinsaurralde can you put benchcmp results?