@matiasinsaurralde 2018-05-18
Rebased this with #1695 changes, `benchcmp` output looks like this, running with `go test -benchmem -run=^$ . -bench ^BenchmarkRegexExtractor$ -benchtime=10s`:
```
% benchcmp old new
benchmark                                  old ns/op     new ns/op     delta
BenchmarkRegexExtractor/HeaderSource-8     185811        141028        -24.10%
BenchmarkRegexExtractor/BodySource-8       180877        2369          -98.69%
BenchmarkRegexExtractor/FormSource-8       191935        150725        -21.47%

benchmark                                  old allocs     new allocs     delta
BenchmarkRegexExtractor/HeaderSource-8     117            51             -56.41%
BenchmarkRegexExtractor/BodySource-8       124            18             -85.48%
BenchmarkRegexExtractor/FormSource-8       145            79             -45.52%

benchmark                                  old bytes     new bytes     delta
BenchmarkRegexExtractor/HeaderSource-8     48704         5206          -89.31%
BenchmarkRegexExtractor/BodySource-8       52981         1569          -97.04%
BenchmarkRegexExtractor/FormSource-8       55096         11597         -78.95%
```