@dencoded 2018-05-18
benchcmp results looks cool!