@matiasinsaurralde 2018-05-18
@dencoded Broke the tests during the rebase, these are the actual results (not bad):
```
% benchcmp old new
benchmark                                  old ns/op     new ns/op     delta
BenchmarkRegexExtractor/HeaderSource-8     177926        144125        -19.00%
BenchmarkRegexExtractor/BodySource-8       178634        143564        -19.63%
BenchmarkRegexExtractor/FormSource-8       187320        150011        -19.92%

benchmark                                  old allocs     new allocs     delta
BenchmarkRegexExtractor/HeaderSource-8     122            57             -53.28%
BenchmarkRegexExtractor/BodySource-8       129            66             -48.84%
BenchmarkRegexExtractor/FormSource-8       150            85             -43.33%

benchmark                                  old bytes     new bytes     delta
BenchmarkRegexExtractor/HeaderSource-8     48942         5446          -88.87%
BenchmarkRegexExtractor/BodySource-8       53216         9786          -81.61%
BenchmarkRegexExtractor/FormSource-8       55331         11836         -78.61%
```